RPN method works
